Simple correction on DB connection.
Basically, when we subscribe to more than a single queue, it was
creating a new connection pool for every new subscription. This is
resource wasteful.

On the other hand, we don't want to lose the ability to re-connect
to another DB when we define a new subscription fn. So, we cache
the connection, but only when we define a new subscription fn with
subscribe-with.